Enable debug toolbar auto-insert when in Docker
